fix - correctly check

+++ app/Http/Controllers/Home/LoginController.php
@@ -34,7 +34,7 @@ class LoginController extends BaseController
 
             // check suspension status
             $ban = $authUser->ban;
-            if ($ban->until->timestamp <= time()) {
+            if ($ban !== null && $ban->until->timestamp <= time()) {
                 $ban->delete();
             }
